Added complete scoreboard functionality and made item text face player

+++ Assets/Scripts/Configuration.cs
@@ -6,7 +6,7 @@ public class Configuration : MonoBehaviour {
 
 
 	public static float presentationTime=3f;
-	public static float maxObjects =3;
+	public static int maxObjects =3;
 
 	public static float minRetrievalDistance = 0.1f;
 	public static float minResponseDistance = 0.3f;
